requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
PySide2==5.15.2
python-vlc==3.0.11115
pywin32==300
requests==2.25.1
shiboken2==5.15.2
urllib3==1.26.2
